requirements.txt
fastapi
uvicorn
python-multipart
openai
python-dotenv
todoist-api-python
sentence-transformers
dateparser

# Google API
google-api-python-client

# Google Auth
google-auth-oauthlib
google-auth-httplib2

# RAG Dependencies
numpy
pydantic
pinecone
scikit-learn
